read file instead of whole dir

+++ display.sh
@@ -1,17 +1,18 @@
 #!/bin/sh
 
-CURRENT_DIR=`dirname "$BASH_SOURCE"`
+CURRENT_DIR="$HOME/.config/cmus/cmus-cover-art"
 COVERS_DIR="$CURRENT_DIR/.cover"
 
 PREVIOUS=""
 CURRENT=""
 
-IMAGE_VIEWER="$CURRENT_DIR/imgcat"
+IMAGE_VIEWER="$HOME/.local/bin/kitty +kitten icat"
 
 clear
 while (true)
 do
-    CURRENT=$(ls ${COVERS_DIR})
+  if test -f "$COVERS_DIR/current.txt"; then
+    CURRENT=`cat $COVERS_DIR/current.txt`
     if [ "$CURRENT" != "$PREVIOUS" ]
     then 
       pkill -TERM -P $$
@@ -19,12 +20,14 @@ do
       if [ "$CURRENT" != "" ]
       then
         PREVIOUS=$CURRENT
-        $IMAGE_VIEWER $COVERS_DIR/$CURRENT &
+        $IMAGE_VIEWER "$CURRENT" &
       else
         echo "::: NO COVER ART :::" 
         PREVIOUS=""
       fi
-    fi   
-    sleep 1
+    fi
+  fi
+
+  sleep 1
 done
